+++ src/snowcli/cli/containers/compute_pool/commands.py
@@ -3,7 +3,7 @@ from enum import Enum
 import typer
 from snowcli.cli.common.decorators import global_options_with_connection
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.cli.snowpark.compute_pool.manager import ComputePoolManager
+from snowcli.cli.containers.compute_pool.manager import ComputePoolManager
 from snowcli.output.decorators import with_output
 from snowcli.output.types import SingleQueryResult, QueryResult, CommandResult